requirements.txt
Django>=2.1.3, <2.2.0
djangorestframework>=3.9.0, <3.10.0
djangorestframework-jwt==1.11.0
django-rest-witchcraft
pytz==2018.3
alembic
python-dateutil
django-oauth-toolkit
django-cors-middleware